requirements.txt
aiogram~=2.12
environs~=8.0.0
python-dotenv~=0.18.0
SQLAlchemy~=1.3.24
gino~=1.0.1